Refactor: Switch <Queue /> over to framer-motion.
